requirements.txt
asgiref==3.8.1
Django==5.1.7
django-debug-toolbar==5.0.1
django-filter==25.1
django-rest-passwordreset==1.5.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Faker==37.0.0
pillow==11.1.0
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
